# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Get Style APIs",
        "description": "Get Style APIs",
        "contact": {
            "email": "dev@get-style.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://dev.get-style.com/api/",
            "description": "local"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register Customer|Shop|Stylist",
                "operationId": "register",
                "requestBody": {
                    "description": "Register a new user.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "description": "fullname of the user.",
                                        "type": "string",
                                        "example": "Asad Shareef"
                                    },
                                    "email": {
                                        "description": "email of the user.",
                                        "type": "string",
                                        "example": "asad@gmail.com"
                                    },
                                    "phone": {
                                        "description": "Phone of the user.",
                                        "type": "string",
                                        "example": "923482269069"
                                    },
                                    "account_type": {
                                        "description": "Type of user.",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "password": {
                                        "description": "Password of the user.",
                                        "type": "string",
                                        "example": "IamAsecretHero"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirmed password of the user.",
                                        "type": "string",
                                        "example": "IamAsecretHero"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login User",
                "operationId": "login",
                "requestBody": {
                    "description": "Login User to the system.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "user login email address provided on signup to Nonsulin",
                                        "type": "string",
                                        "example": "asad@getstyle.com"
                                    },
                                    "password": {
                                        "description": "user secure password.",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout User",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "user_access_token": []
                    }
                ]
            }
        },
        "/user/change-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change password",
                "operationId": "changePassword",
                "requestBody": {
                    "description": "Change Password.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "old_password": {
                                        "description": "user old password.",
                                        "type": "string",
                                        "example": "Test123!"
                                    },
                                    "password": {
                                        "description": "user new password.",
                                        "type": "string",
                                        "example": "Test123!"
                                    },
                                    "password_confirmation": {
                                        "description": "user new password confirmation.",
                                        "type": "string",
                                        "example": "Test123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "user_access_token": []
                    }
                ]
            }
        },
        "/metadata": {
            "get": {
                "tags": [
                    "Meta Configuration"
                ],
                "summary": "Get App Metadata",
                "operationId": "getWebAppMetaData",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/stylist/top-rated": {
            "get": {
                "tags": [
                    "Stylist"
                ],
                "summary": "Top Rated Stylists",
                "operationId": "getTopRatedStylists",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "user_access_token": []
                    }
                ]
            }
        },
        "/stylist/{offset}": {
            "get": {
                "tags": [
                    "Stylist"
                ],
                "summary": "Get all stylists",
                "operationId": "getAllStylists",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "area",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "shop_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "services[]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "user_access_token": []
                    }
                ]
            }
        },
        "/user/profile-picture": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Store profile pic for logged in User.",
                "operationId": "updateProfilePicture",
                "requestBody": {
                    "description": "Store profile pic for logged in User.",
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "profile_picture": {
                                        "description": "select profile picture for logged in User.",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "user_access_token": []
                    }
                ]
            }
        },
        "/profile/update": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Edit patient profile data",
                "operationId": "editProfile",
                "requestBody": {
                    "description": "Edit patient profile data.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "description": "name",
                                        "type": "string",
                                        "example": "Full name"
                                    },
                                    "email": {
                                        "description": "user email",
                                        "type": "string",
                                        "example": "abc@gg.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "user_access_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "user_access_token": {
                "type": "apiKey",
                "description": "user access token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}